[LARGE] Remove Eugenia
The GMF Tooling project was terminated due to a lack of development
activity:

https://gitlab.eclipse.org/eclipsefdn/emo-team/emo/-/issues/593

As part of the termination process, the GMF Tooling update site was
deleted: this broke our target platform, and meant we would not be able
to build Eugenia anymore. This commit removes all Eugenia components
and tests from the repository, to restore the build.