Add new strategy for Maven dependencies
The "MavenDependencies" element in the target platform can be extended
to pull approved dependencies directly from Maven Central. Then, these
dependencies can be added as "wrapped" plugins in those projects that
need them.

Please note: to allow Eclipse finding these Maven dependencies during the
installation of features, the wrapped plugins also need to be
added to those features that contain one or more projects using the dependencies.

This commit includes as example how to use PlantUML in the Picto
feature.