eol.dap: support approximate suffix-based matching of breakpoints
There are cases where the URLs reported by the module are different than
those sent in DAP requests. This happens, for instance, if we're running
an Epsilon script from the classpath of a Java program, instead of
running it directly from a file in the filesystem.

This commit introduces an approach which will match based on the longest
matching sequence of trailing URI components. Suppose we had set a
breakpoint on this file in our IDE:

file:/your/path/to/sources/a.b.c/epsilon/main/a.eol

If we had modules for these:

platform:/plugin/a.b.c/epsilon/main/a.eol
platform:/plugin/a.b.c/epsilon/util/a.eol

Then it would prefer the main/a.eol option, as it would match a longer
sequence of trailing URI components.