examples.eol.dap: simplify 1-click debug from VS Code
VS Code can do the waiting for the server for us, so long as we run the
Gradle task as a background task, and provide a problemMatcher with the
"background" key set up to detect when we are ready to debug.

This requires the latest version of the remote-debug branch of the
Epsilon extension for VS Code, which provides a problemMatcher with the
necessary "background" patterns.