Move breakpoint verification to IEolDebugger
This is needed in order to have language-specific behaviour for
breakpoint verification. For instance, in EGX we won't be able to
immediately verify EGL breakpoints as the exact EGL templates to be run
will not be known until we run the program.

With this new design, we can make EGX leave non-EGX breakpoints as
"pending", to be resolved later.