[EGX] Make EGL scripts reuse execution controller of EGX module
We need to have the EGL scripts use the same execution controller as the
EGX module that started them. This completes the first round of fixes
needed for having breakpoints in EGL files work from EGX.

There is still some pending work: getting breakpoints working on EGL
files that are being loaded from the classpath will require populating
the breakpoints at the time an EGL script is loaded, rather than at the
time of starting the execution of the EGX script.