EolDebugger: treat each controlled module execution as its own thread
This handles the case where we have an IEolModule execute() another
IEolModule (so long as they all use the same execution controller). It
is useful for EGX when it invokes EGL scripts: the EGX script will be
treated as one thread for debugging, and the EGL script will be treated
as a different thread.