[DAP] Support breakpoints in EGL scripts run from classpath-based EGX
This commit changes the design of the debug adapter a bit more: instead
of having the same debug adapter as the debug target for all *Debugger
objects, each *Debugger object gets its own, with its own set of
resolved breakpoints. This allows for honouring the language-specific
logic of each *Debugger object, and makes it easier to throw away stale
state once we finish running a submodule.