[DAP] Unify current location resolution across stackTrace() and sendOutput
When we had a crash in an ECL executable block that only had one
expression, trying to send out the error while we didn't have a current
location would crash. We already had logic for this situation in
stackTrace(), but it wasn't getting reused here.

This commit extracts the logic to a new method, which is then reused
both for reporting stack traces, and for reporting the location that
produced a given line of output.