eml.dt: redesign on top of a 'combo module' for debugging
Due to how LSP4E works, we cannot use the old approach in the EML launch
configurations where we run the ECL delegate then the EML delegate: this
confuses LSP4E while debugging, as it considers that the debuggee for
this launch has completed.

Instead, we use a new EclEmlModule that acts as an orchestration layer,
running the ECL module and then the EML module, in a similar way to how
EgxModule orchestrates multiple modules.