Rewrite module equivalence tests
The old tests created and ran all modules in one go, and then checked things.
This rewrite instead runs modules as part of the test and ensures resources
are freed before the next test. This improves isolation between tests, and
makes it easier to see when things go wrong.

The old tests also depended on the alphabetical ordering of the test method
names. This is no longer the case for the new version.