[DAP] Expand synchronized block to cover sending of stopped message
When running tests from Github Actions, there was a race condition between the
code handling the "stopped" message and the code that was trying to continue the
execution after the end of the test. Sometimes, the test would try to release the
"suspendLatch" before it was set, which would result in the "continue" request
effectively being ignored by the debug adapter.

This commit extends the synchronized block in the suspend() method so any code
trying to call resumeAllThreads() while reacting to the "stopped" message will
have to wait until the suspendLatch has been enabled and the execution of the
Epsilon script has been suspended by having the Thread wait().