[Fixes #118] Avoid spurious ambiguity warning due to subpackages
Suppose we had a metamodel like this:

package A {
  package B {
    class C {}
  }
}

If someone mentioned both A and B in the package registry of an EMF model,
the current code would find two ways to get to C (A -> B -> C, and B -> C),
and would issue a spurious "type is ambiguous" warning even through both
ways lead to the same "C" type.

This commit tweaks the ambiguity detection logic to avoid producing this
spurious warning, and adds a test for it.