[DAP] Add support for conditional breakpoints (fixes #121)
We support the DAP API for conditional breakpoints now. This works by
creating a throwaway EolModule with a copy of the existing context,
but with a different instance of the executor factory (as we do not
want to debug the expression being evaluated), and a separate copy of
the frame stack.

At the moment, this is only usable from VS Code. It appears that we
would need to write some additional UI code to have it work from Eclipse:
we need to add an edit view for Epsilon breakpoints for the condition,
and LSP4E needs to pass on this information via DAP.