Avoid using Hamcrest containsString
It seems that with the target platform set to Eclipse 2024-12, the
test.acceptance project is using the old copy of Hamcrest that is in
JUnit 4, instead of Hamcrest 3.0, so we don't have access to containsString.

Ideally we'd like to move to JUnit 5, which drops its embedded copy of
Hamcrest altogether, but that would require rewriting many tests, so
we can hold off on this until we start work on 2.7.

Alternatively, we could stop using the centralised test.dependencies
project and start to move tests gradually to JUnit 5. That is for later,
though.