[DAP] Refuse to evaluate expressions unless stopped at a breakpoint
LSP4E does not seem to take into account whether the program is stopped
at a breakpoint or not for its evaluate() requests, so we need to handle
this ourselves. The DAP server will now reply with "(pending)" if an
evaluate() request is received while the program is not stopped at a
breakpoint. Otherwise, we may be interfering with its normal execution.