[DAP] Create a dedicated ReusableEpsilonDebugServer class + switch to polling
Indefinitely blocking on the execution queue was hard to use as a user could
forget to enqueue a module right after sending the termination request, which
would make the program never terminate.

Instead, we now have a configurable poll interval on the blocking queue, which
is set to 200ms - frequent enough to avoid excessive delays, but not so frequent
that it'll result in busy waits.

I have also changed the ExecutionQueueModule so it knows about the debug adapter
it is being used with - it will ask the debug adapter to attach to any modules it
runs in advance, so now we can assume that the *Debugger class will always be
used to control their execution. This means we can revert the workaround we had
created to solve #167 in EgxDebugger.

With these tweaks, we can provide a ReusableEpsilonDebugServer class which would
be much easier to reuse downstream, and from our Ant workflow tasks. This is a
specialised debug server that will automatically set up the ExecutionQueueModule
accordingly.