[DAP] Fix prefix URI-to-path mappings on Windows (#175)
Besides having issues on WIndows, the old code required being careful to use trailing slashes for any URIs
that mapped to directories. This is not needed anymore, as we simply check whether the Path used at the
other end of a mapping resolves to a file or to a directory.