Reorganize the standalone example
The example was halfway Mavenized but not all the way, and some
classes were using the deprecated base class.

This commit reorganizes the example so that it is fully Mavenized
(sources in src/main/java, resources in src/main/resources), and
removes the deprecated base class after revamping the examples
that were using it.

Epsilon programs and metamodels are now resources which would be
bundled with the rest of the program, whereas models are no longer
bundled with the program.