Delete subtemplate traceability tests
Subtemplates contributing to the trace of their parent was to some
extent incidental due to the previous implementation of fine
traceability. Changing the "egl" string of
SubtemplateShouldContributeToTraceOfParent for example to
[%=TemplateFactory.prepare(openOutputTag + 'EClass.all.first.name' +
closeTag + "Static Text").process()%] produced no trace. Using
TemplateFactory to execute sub-templates is no longer recommended
practice after the introduction of EGX and @template operations so I'm
deleting these tests as they correctly fail with the new fine
traceability implementation.